requirements.txt
# ARCANA AI — Dependencies
# Python 3.11+

# Core
httpx>=0.27
pydantic>=2.0
python-dotenv>=1.0

# LLM (via OpenRouter)
openai>=1.0

# X / Twitter
tweepy>=4.14

# Payments
stripe>=8.0

# Notifications
discord-webhook>=1.3
python-telegram-bot>=21.0
